requirements.txt
asgiref==3.2.7
certifi==2020.4.5.2
chardet==3.0.4
Django==3.0.7
django-crispy-forms==1.9.1
idna==2.9
lxml==4.5.1
pytz==2020.1
requests==2.23.0
soupsieve==2.0.1
sqlparse==0.3.1
urllib3==1.25.9